Merges origin's two commits (real build-error fixes + Inter font + screenshot
testing, and the calm/premium redesign with 3D hero, bottom nav bar, and
richer onboarding) with the local Nutrition & Fitness module work.

Conflict resolution:
- ProfileData.kt: kept both sides' new UserProfileEntity fields (remote's
  immersiveModeEnabled/smokingStatus/cannabisUseFrequency/alcoholUseFrequency
  plus this session's ageYears/dailyWaterGoalMl/dailyCalorieGoalOverride).
  Standardized on remote's heightCm: Int (already wired through their
  build-tested onboarding flow) instead of this session's Float. Kept
  remote's 4-value ActivityLevel enum (used by onboarding's chip UI) and
  added the tdeeMultiplier property it needs for NutritionGoalCalculator,
  rather than reintroducing a 5th MODERATE value that would have required
  touching onboarding UI code that was already working.
- DashboardScreen.kt: kept both new icon imports (Restaurant for the
  Nutrition module card, Notifications for remote's redesigned settings
  entry point).
- strings.xml (en/it): merged the Nutrition strings with remote's new
  History Hub / bottom-tab strings; dropped a duplicate activity_level_*
  block that would have collided with remote's earlier definitions of the
  same keys.

Verified after merge: no leftover conflict markers, brace/paren balance
across all files, full en/it string-key parity with no duplicates, and
every ViewModel's AppContainer property reference matches an actual
AppContainer member.

+++ app/src/main/java/com/dopamind/app/feature/nutrition/domain/NutritionGoalCalculator.kt
@@ -30,7 +30,7 @@ object NutritionGoalCalculator {
     private const val CALORIES_PER_GRAM_CARBS = 4
     private const val CALORIES_PER_GRAM_FAT = 9
 
-    fun calculateBmr(weightKg: Float, heightCm: Float, ageYears: Int, sex: BiologicalSex): Double {
+    fun calculateBmr(weightKg: Float, heightCm: Int, ageYears: Int, sex: BiologicalSex): Double {
         val base = 10 * weightKg + 6.25 * heightCm - 5 * ageYears
         return when (sex) {
             BiologicalSex.MALE -> base + 5
@@ -43,7 +43,7 @@ object NutritionGoalCalculator {
 
     fun calculateGoal(
         weightKg: Float,
-        heightCm: Float,
+        heightCm: Int,
         ageYears: Int,
         sex: BiologicalSex,
         activityLevel: ActivityLevel,
